+++ testgame/asteroid.hpp
@@ -8,8 +8,8 @@
 class Asteroid : public jimp::AnimatedSprite {
   
 private:
-    static const int FORCE;
-    static const int ROTATION_DEGREES_PER_SECOND;
+    static const uint16_t FORCE;
+    static const uint16_t ROTATION_DEGREES_PER_SECOND;
     static const float HIT_ANIMATION_DURATION_IN_SECONDS;
     float directionAngle;
     float totalHitAnimationDuration = 0.0F;
