+++ testgame/asteroidSpawner.hpp
@@ -7,7 +7,7 @@
 class AsteroidSpawner {
     
 private:
-    static const int SECONDS_BETWEEN_ASTEROID_SPAWN;
+    static const uint8_t SECONDS_BETWEEN_ASTEROID_SPAWN;
     jimp::GameEngine* gameEngine = nullptr;
     std::list<Asteroid*>* asteroids = nullptr;
     float timeSinceLastAsteroid = 0.0F;
